{"schema":"decodex/commit/1","summary":"Split add_note validation non-English traversal.","authority":"manual"}
